refactor : 레이싱카 이동(RacingCarMovement) package move
